Add `Polygon::contains_triangle`
